+++ bench/coworker/src/main/kotlin/io/kungfury/coworkerperf/EchoJavaJob.java
@@ -1,11 +1,12 @@
 package io.kungfury.coworkerperf;
 
 import io.kungfury.coworker.BackgroundJavaWork;
+import io.kungfury.coworker.WorkGarbage;
 import io.kungfury.coworker.dbs.ConnectionManager;
 
 public class EchoJavaJob extends BackgroundJavaWork {
-    public EchoJavaJob(ConnectionManager _connManager, long id, int stage, String strand, int priority) {
-        super(id, stage, strand, priority);
+    public EchoJavaJob(ConnectionManager _connManager, WorkGarbage garbageHeap, long id, int stage, String strand, int priority) {
+        super(garbageHeap, id, stage, strand, priority);
     }
 
     @Override