+++ bench/coworker/src/main/kotlin/io/kungfury/coworkerperf/EchoJob.kt
@@ -2,17 +2,19 @@ package io.kungfury.coworkerperf
 
 import io.kungfury.coworker.dbs.ConnectionManager
 import io.kungfury.coworker.BackgroundKotlinWork
+import io.kungfury.coworker.WorkGarbage
 
 /**
  * A job that just echos out.
  */
 class EchoJob(
     private val connectionManager: ConnectionManager,
+    garbageHeap: WorkGarbage,
     id: Long,
     stage: Int,
     strand: String,
     priority: Int
-) : BackgroundKotlinWork(id, stage, strand, priority) {
+) : BackgroundKotlinWork(garbageHeap, id, stage, strand, priority) {
     override fun serializeState(): String = ""
 
     override suspend fun Work(state: String) {