+++ bench/coworker/src/main/kotlin/io/kungfury/coworkerperf/EmptyJob.kt
@@ -1,6 +1,7 @@
 package io.kungfury.coworkerperf
 
 import io.kungfury.coworker.BackgroundKotlinWork
+import io.kungfury.coworker.WorkGarbage
 import io.kungfury.coworker.dbs.ConnectionManager
 
 /**
@@ -8,11 +9,12 @@ import io.kungfury.coworker.dbs.ConnectionManager
  */
 class EmptyJob(
     private val connectionManager: ConnectionManager,
+    garbageHeap: WorkGarbage,
     id: Long,
     stage: Int,
     strand: String,
     priority: Int
-) : BackgroundKotlinWork(id, stage, strand, priority) {
+) : BackgroundKotlinWork(garbageHeap, id, stage, strand, priority) {
     override fun serializeState(): String = ""
 
     override suspend fun Work(state: String) {